+++ FoodWare/View/UserControls/UC_Configuracion.cs
@@ -2,17 +2,17 @@
 using System.Drawing;
 using System.Windows.Forms;
 
-namespace FoodWare.UserControls
+namespace FoodWare.View.UserControls
 {
-    public partial class UC_Ventas : UserControl
+    public partial class UC_Configuracion : UserControl
     {
-        public UC_Ventas()
+        public UC_Configuracion()
         {
             InitializeComponent();
 
             var lbl = new Label
             {
-                Text = "Ventas",
+                Text = "Configuracion",
                 Dock = DockStyle.Fill,
                 TextAlign = ContentAlignment.MiddleCenter,
                 Font = new Font("Segoe UI", 14, FontStyle.Bold)
