+++ FoodWare/View/UserControls/UC_Empleados.cs
@@ -2,17 +2,17 @@
 using System.Drawing;
 using System.Windows.Forms;
 
-namespace FoodWare.UserControls
+namespace FoodWare.View.UserControls
 {
-    public partial class UC_Reportes : UserControl
+    public partial class UC_Empleados : UserControl
     {
-        public UC_Reportes()
+        public UC_Empleados()
         {
             InitializeComponent();
 
             var lbl = new Label
             {
-                Text = "Reportes",
+                Text = "Empleados",
                 Dock = DockStyle.Fill,
                 TextAlign = ContentAlignment.MiddleCenter,
                 Font = new Font("Segoe UI", 14, FontStyle.Bold)
