+++ FoodWare/View/UserControls/UC_Finanzas.cs
@@ -2,17 +2,17 @@
 using System.Drawing;
 using System.Windows.Forms;
 
-namespace FoodWare.UserControls
+namespace FoodWare.View.UserControls
 {
-    public partial class UC_Configuracion : UserControl
+    public partial class UC_Finanzas : UserControl
     {
-        public UC_Configuracion()
+        public UC_Finanzas()
         {
             InitializeComponent();
 
             var lbl = new Label
             {
-                Text = "Configuracion",
+                Text = "Finanzas",
                 Dock = DockStyle.Fill,
                 TextAlign = ContentAlignment.MiddleCenter,
                 Font = new Font("Segoe UI", 14, FontStyle.Bold)
