+++ FoodWare/View/UserControls/UC_Ventas.resx
